test(extra): relocate CLI argument-parsing tests into testsuite-extra
ironrdp-agent sets [lib] test = false, so cli.rs's inline #[cfg(test)]
tests never ran under cargo test --workspace. Move the five tests that
exercise only the public Cli surface (clap's Parser/CommandFactory
traits, no private fields or enum variants) into
ironrdp-testsuite-extra/tests/agent.rs, alongside the existing
remote_exit_status coverage for the same module.

The sandbox.rs and sandbox_grpc.rs test modules stay inline: both are
declared pub(crate) in lib.rs, so every item inside them is
unreachable from an external crate regardless of its own visibility.
The remaining cli.rs tests stay inline for the same reason, touching
private Command/SandboxCommand enum variants or private helper
functions.
